fix: cleanup from full codebase review
foxess_grid_charge_scheduler.py:
  - _update_charge_state: evaluate each window independently for skip
    eligibility — mixed targets (morning=100, evening=85) now handled
    correctly instead of falling through to clear_skip()

scheduler_models.py:
  - ProximityResult.radiation: int | None → float | None (matches
    actual return type of get_solar_forecast)

foxess_api.py:
  - Add comments explaining why notify_warning fires per failure
    (acceptable: 3-retry backoff, failures rare in practice)

weather.py:
  - Revert timezone Europe/Warsaw → auto (Open-Meteo infers from
    lat/lon; hardcoded value was a temporary workaround during outage)

README.md:
  - Update dynamic window description to cover both window 1 and 2

+++ scheduler_models.py
@@ -29,6 +29,6 @@ class ChargePlan:
 @dataclass
 class ProximityResult:
     should_run: bool
-    radiation: int | None = None
+    radiation: float | None = None
     low_solar: bool | None = None
     skip_reason: str | None = None
